+++ localdocs/__init__.py
@@ -2,10 +2,13 @@
 
 from localdocs.chunker import chunk_blocks
 from localdocs.config import LocalDocsConfig, load_config
+from localdocs.flashcards import export_anki_tsv, generate_flashcards
 from localdocs.indexer import LocalIndex, build_index
+from localdocs.obsidian import export_obsidian_vault
 from localdocs.parser import parse_path, parse_uploaded_file
 from localdocs.qa import answer_question
 from localdocs.search import search
+from localdocs.study import export_study_questions_markdown, generate_study_questions
 from localdocs.summarizer import summarize_documents
 
 __all__ = [
@@ -14,6 +17,11 @@ __all__ = [
     "answer_question",
     "build_index",
     "chunk_blocks",
+    "export_anki_tsv",
+    "export_obsidian_vault",
+    "export_study_questions_markdown",
+    "generate_flashcards",
+    "generate_study_questions",
     "load_config",
     "parse_path",
     "parse_uploaded_file",